backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["stephen <kernan.software@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg2 = "^2.9.7"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
drf-yasg = "^1.21.7"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "api.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "tests/*.py"]